requirements.txt
openai>=1.30
jsonschema>=4.0
python-dotenv>=1.0
